% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/advproginR.R
\docType{package}
\name{advproginR}
\alias{advproginR}
\alias{advproginR-package}
\title{advproginR package implements the Euclidean gcd algorithm and the Dijkstra algorithm}
\description{
The advproginR package contains two functions and one dataset.
}
\section{advproginR functions}{

The advproginR functions \code{euclidean} and \code{dijkstra}
compute the greatest common denominator of two integers and
the shortest distance to every node in a weighted graph, respectively.
}

\section{advproginR data}{

The advproginR package includes the dataset \code{wiki_graph}
which is a data frame which represents the graph found on the wikipedia
page for Djikstras algorithm.
}

