% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{Implements the dijkstra algorithm}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{graph}{A data.frame object with three columns, representing a graph
  with weighted edges. the column content are as follows:


  \describe{

  \item{1}{The integer numbers of the starting nodes of the edges in the
  graph}

  \item{2}{The integer numbers of the ending nodes of the edges in the graph}

  \item{3}{The numeric edge weights of the edges in the graph}

  }

  Each edge is considered to have two starting nodes and thus is listed two
  times. This is required.}

\item{init_node}{The integer of the starting node from which dijkstras
  algorithm proceeds to find the shortest distances to the other nodes.}
}
\value{
\code{resultvec2} A numeric vector with the shortest distances to the
  other nodes, ordered according to the numbering of the nodes.
}
\description{
\code{dijkstra} returns the shortest distance from the inital node in a graph
to all other nodes.
}
\details{
Dijkstras algorithm is implemented in function \code{dijkstra} which takes
first argument \code{graph} a three column data frame which represents a
weighted graph and second argument \code{init_node} an integer to represent
the initial node. It then proceeds to produce the shortest distances from an
initial node to the rest of the nodes in a given graph. It returns a vector
with the shortest distances in the order of the node numbering.
}
\examples{
crosssquare <- data.frame(c(1,1,1,2,2,2,3,3,3,4,4,4),
c(2,3,4,3,4,1,4,1,2,1,2,3),c(1,6,2,5,3,1,2,6,5,2,3,2))
start <- 3
dijkstra(crosssquare, start)
}
\references{
\url{http://en.wikipedia.org/wiki/Dijkstra\%27s_algorithm}
}

