% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Implements the Euclidean algorithm}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{Should be a non negative integer.}

\item{b}{Should also be a non negative integer.}
}
\value{
Returns the greatest common divisor between \code{a} and \code{b}.
}
\description{
\code{euclidean} returns the greatest common divisor of two integer numbers.
}
\details{
\code{euclidean} takes two integer numbers, \code{a} and \code{b}, and it
divides the largest with the smallest one. Then the function takes the
remainder and divides the smallest number between \code{a} and \code{b} with
the remainder. This process continues until the function encounters a
division where the remainder is zero and it returns the last denominator of
the division.
}
\examples{
euclidean(5,18) returns 1
}
\references{
\url{http://en.wikipedia.org/wiki/Euclidean_algorithm}
}

