% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wiki_graph.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{The graph found on the wikipedia page for Dijkstras algorithm}
\format{A dataframe with 18 observations on the following 3 variables.

  \describe{

  \item{v1}{The numeric index number of the starting node of the 18 edges in the
  graph}

  \item{v2}{The numeric index number of the ending node of the 18 edges in the
  graph}

  \item{w}{The numeric edge weights of the 18 edges in the graph}

  }}
\source{
\url{http://en.wikipedia.org/wiki/Dijkstra\%27s_algorithm}
}
\usage{
wiki_graph
}
\description{
The advproginR package includes the dataset \code{wiki_graph} which is a data
frame which represents the graph found on the wikipedia page for Djikstras
algorithm. Each edge is represented two times, one time for each direction
}
\keyword{datasets}

